feat(gui): add support for dropping files into the file area
