+++ src/neosca/ns_io.py
@@ -263,6 +263,24 @@ class Ns_IO(metaclass=Ns_IO_Meta):
             stem = f"{stem} ({occurrence})"
         return stem
 
+    @classmethod
+    def find_files(cls, folder_path: str, is_recursive: bool) -> list[str]:
+        file_paths: list[str] = []
+        if is_recursive:
+            file_paths.extend(
+                os_path.join(dir_path, file_name)
+                for dir_path, _, file_names in os.walk(folder_path)
+                for file_name in file_names
+                if not file_name.startswith(Ns_IO.HIDDEN_PREFIXES)
+            )
+        else:
+            file_paths.extend(
+                os_path.join(folder_path, file_name)
+                for file_name in next(os.walk(folder_path))[2]
+                if not file_name.startswith(Ns_IO.HIDDEN_PREFIXES)
+            )
+        return file_paths
+
 
 class Ns_Cache:
     CACHE_EXTENSION = ".pickle.lzma"
